+++ src/graderservice/graderservice/models.py
@@ -1,17 +1,6 @@
-#  (C) Copyright IllumiDesk, LLC, 2020.
-#
-#  Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
-#  the License. You may obtain a copy of the License at
+from flask_sqlalchemy import SQLAlchemy
 
-#  http://www.apache.org/licenses/LICENSE-2.0
-
-#  Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
-#  an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
-#  specific language governing permissions and limitations under the License.
-
-import flask_sqlalchemy
-
-db = flask_sqlalchemy.SQLAlchemy()
+db = SQLAlchemy()
 
 
 class GraderService(db.Model):
@@ -29,7 +18,7 @@ class GraderService(db.Model):
         The grader Service object's name and url properties
     """
 
-    __tablename__ = 'grader_services'
+    __tablename__ = "grader_services"
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(60), unique=True, nullable=False)
     course_id = db.Column(db.String(50), nullable=False)
@@ -38,7 +27,9 @@ class GraderService(db.Model):
     admin = db.Column(db.Boolean, default=True)
     api_token = db.Column(db.String(150), nullable=True)
 
-    db.create_all()
+    def get_id(self):
+        """Return the service ID as a unicode string (`str`)."""
+        return str(self.id)
 
     def __repr__(self):
         return "<Service name: {} at {}>".format(self.name, self.url)